pyproject.toml
[tool.poetry]
name = "leporem-art"
version = "0.1.0"
description = "backend for leporem art"
authors = ["jshong0907 <jshong0907@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10.0"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
sentry-sdk = "^1.26.0"
mangum = "^0.17.0"
django-s3-storage = "^0.14.0"
requests = "^2.31.0"
pymysql = "^1.1.0"
django-extensions = "^3.2.3"
cryptography = "^41.0.1"
boto3 = "^1.28.1"
django-cors-headers = "^4.1.0"
django-storages = "^1.13.2"
pillow = "^10.0.0"
paginator = "^0.5.1"
pyjwt = "^2.7.0"
pre-commit = "^3.3.3"
django-filter = "^23.2"
social-auth-app-django = "^5.2.0"
social-auth-core = "^4.4.2"
dj-rest-auth = "^4.0.1"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
django-extensions = "^3.2.3"
notebook = "^6.5.4"
pre-commit = "^3.3.3"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
    \.git
    | ^.*\b(migrations)\b.*$
    | ^.*\b(settings)\b.*$
'''
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
skip = [".git", "migrations"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'leporem_art.settings.test'
python_files = ['tests/*.py']
adopts = '--create-db'
